requirements.txt
azure-storage==0.20.0
azure==2.0.0
dj-database-url==0.4.2
Django==1.10.3
django-modelcluster==3.1
django-taggit==0.22.1
django-treebeard==4.1.2
django-storages==1.6.5
djangorestframework==3.6.2
django-wysiwyg-redactor==0.5.1
gunicorn==19.7.1
oauthlib==2.0.2
Pillow==4.1.0
python-dateutil==2.6.0
python-decouple==3.1
python-memcached==1.59
psycopg2==2.7.3.2
pyvirtualdisplay==0.2.1
requests==2.18.4
requests-file==1.4.1
requests-oauthlib==0.8.0
selenium==3.9.0
simplejson==3.10.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
Unidecode==0.4.21
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.0
